// packages/schema/persistence.d.ts
export type Name = string;
export type Projectid = string;
export type Revision = number;
// ...
export interface ProjectSummary {
  name: Name;
  projectId: Projectid;
  revision: Revision;
}